Fix #10386. Mistake made while refactoring (#10607)
Added some additional comments whilst investigating the issue. Ultimate fix should look at changing the bounding box units but unsure of consequences

+++ src/openrct2/paint/Paint.cpp
@@ -890,7 +890,7 @@ paint_struct* sub_98197C(
 
     session->LastRootPS = ps;
 
-    auto attach = CoordsXY{ ps->bounds.x, ps->bounds.y }.Rotate(session->CurrentRotation);
+    auto attach = CoordsXY{ (int16_t)ps->bounds.x, (int16_t)ps->bounds.y }.Rotate(session->CurrentRotation);
     switch (session->CurrentRotation)
     {
         case 0:
